Week 2: Fix Code model
